+++ backend/app/v1/cart/schemas.py
@@ -1,0 +1,11 @@
+from sqlmodel import SQLModel
+from uuid import UUID
+
+
+class CartItemCreate(SQLModel):
+    harvest_id: UUID
+    qty: int
+
+
+class CartItemUpdate(SQLModel):
+    qty: int | None = None
