+++ backend/app/v1/consumer/responseModels.py
@@ -15,22 +15,3 @@ class HTTPExceptionResponse(SQLModel):
     detail: str
 
     model_config = {"from_attributes": True}
-
-
-class ShowCartItem(SQLModel):
-    id: UUID
-    cart_id: UUID
-    harvest_id: UUID
-    qty: int
-    rate: float
-
-    model_config = {"from_attributes": True}
-
-
-class ShowCart(SQLModel):
-    id: UUID
-    consumer_phone_no: int
-    total_amount: float
-    cart_items: list[ShowCartItem] = []
-
-    model_config = {"from_attributes": True}
