+++ backend/app/v1/consumer/schemas.py
@@ -5,12 +5,3 @@ from uuid import UUID
 class ConsumerUpdate(SQLModel):
     first_name: str | None = None
     last_name: str | None = None
-
-
-class CartItemCreate(SQLModel):
-    harvest_id: UUID
-    qty: int
-
-
-class CartItemUpdate(SQLModel):
-    qty: int | None = None
